Rename google_api_lib to google_api
- Fix mocks
- Fix tests
